Revise repo to be a base rather than an extension
My strategy before was to have this repo as
an extension to a finished base. However this
does not play well with the way volumes are
persistent in docker.

Therefore I have changed my approach to be
using this as a base and the old repo as
extension.

The base is not ment to be used directly
and therefore does not persist the home
directory.